fixing data type

+++ train.py
@@ -163,7 +163,7 @@ learning_rate = float(args.lr)
 weight_decay = float(args.weight_decay)
 momentum = float(args.momentum)
 multiplier = float(args.multiplier)
-classif = float(args.classif)
+classif = int(args.classif)
 
 char_embed = Char_embedding(char_emb_dim, char_max_len, asc=args.asc, random=True, device=device)
 if args.load or int(args.run) > 1:
